Move common encode code to scapp.h.
